Split ReactNoop into normal and persistent exports
