+++ packages/react-reconciler/src/__tests__/ReactPersistent-test.internal.js
@@ -28,10 +28,8 @@ describe('ReactPersistent', () => {
     ReactPortal = require('shared/ReactPortal');
   });
 
-  const DEFAULT_ROOT_ID = 'persistent-test';
-
   function render(element) {
-    ReactNoopPersistent.renderToPersistentRootWithID(element, DEFAULT_ROOT_ID);
+    ReactNoopPersistent.render(element);
   }
 
   function div(...children) {
@@ -44,7 +42,7 @@ describe('ReactPersistent', () => {
   }
 
   function getChildren() {
-    return ReactNoopPersistent.getChildren(DEFAULT_ROOT_ID);
+    return ReactNoopPersistent.getChildren();
   }
 
   it('can update child nodes of a host instance', () => {
